requirements.txt
twitchio==1.2.3
python-dotenv==0.17.0